requirements.txt
flask
docxtpl
supabase
python-dotenv